+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Add metric derived from string metric in metric report.
+- Add average bitrate as a second criterion for stream selection in priority based downlink policy
 
 ### Fixed
 * Fixed missing videos, or unnecessarily long freezes when switching simulcast streams.